De-monoliticize MappingPane.cpp
While at it, give things proper names. VariantDisplay
displays the Variant (even though the wireframe tab still
says Sample). MappingPane is now MacroMappingVariantPane since
it has all three. etc...

No functional changes. Just moving code around and renaming it
so I can tackle the next set of changes without the code being
a morass

It contained all 3 screens as roughly cpp only classes, which
was untenable. This moves the whole mapping pane top area

f
